requirements.txt
asgiref==3.8.1
Django==5.0.4
django-browser-reload==1.12.1
python-decouple==3.8
sqlparse==0.5.0
pyodbc==5.1.0
mssql-django==1.5
whitenoise==6.7.0
gunicorn==22.0.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
